StageMix: fix 13 bugs found by the adversarial test team
Five agents attacked the engine (chaos scenarios, invariant fuzzing,
musicality review, perception false-positives, code audit) with 2,347
lines of tests against the real code. They hit the session limit
mid-run; I salvaged their test files, reproduced every finding, fixed
the confirmed bugs, and folded regressions into the suite.

CRASHES (would have killed the app mid-show):
- takeover/override with a fader above the +2 dB cap crashed every
  tick (coerceIn with min > max) -> boundOffset() clamps the window
- Meters.decode integer-overflow throw on ~24% of hostile buffers
  (4 + count*2 overflows) -> Long comparison
- absurd override values (+100 dB, NaN) crashed after the hold expired
  -> sanitized and clamped

WRONG BEHAVIOR ON STAGE:
- boost budget starved recovery: a quiet channel climbing back TOWARD
  its baseline was blocked by other channels' boosts, so a buried
  vocal could sit unrescued all night. Budget now governs only the
  part of a move that ends ABOVE baseline
- a long gap (app paused, Wi-Fi outage, clock jump) authorized a
  single 15 dB jump -> dt clamped to ~1 tick
- ducks lingered on channels that had gone silent (into the next song)
  -> release when the channel leaves the band
- a fresh takeover inherited stale override holds/idle flags, leaving
  channels unmanaged for the new act -> clean slate
- a quiet channel JOINING moved the mix mean enough to look like a
  broadband event and froze boosts for 5 s -> membership-aware
- MIX HEALTH reported a flattering 100% vocal-on-top when there was
  no vocal channel at all -> reports n/a until it has samples
- NaN/Inf in one RTA bin propagated into an EQ write to the mixer
  -> sanitized at both meter and RTA entry
- takeover's own enquiry replies were treated as human fader moves
  (the  guard was declared but never set) -> bogus
  overrides on every takeover, now guarded
- 'Hand back the mains' was undone by the next tick -> revert also
  pauses the autopilot
- a howl alternating between adjacent bins evaded the watchdog
  forever -> neighbourhood-tolerant persistence
- singers near the male/female register boundary flapped the doctor
  blind every frame -> 6-frame debounce

Suite: 79 tests (13 new regressions + the chaos team's 10 scenario
tests: musician swaps, 20-min singer rotation, everyone-plugs-in-at-
once, dead air, flickering cables, 12 random acts back to back).

+++ stagemix/engine/src/main/kotlin/com/stagemix/engine/Meters.kt
@@ -22,7 +22,9 @@ object Meters {
         if (blob.size < 4) return null
         val buf = ByteBuffer.wrap(blob).order(ByteOrder.LITTLE_ENDIAN)
         val count = buf.int
-        if (count < 0 || blob.size < 4 + count * 2) return null
+        // count * 2 overflows for hostile/huge values -> compare in Long
+        if (count < 0 || blob.size.toLong() < 4L + count.toLong() * 2L)
+            return null
         return FloatArray(count) { buf.short / 256f }
     }
 }
